modules/figure-panel-detection/requirements-cpu.txt
nb_utils
loguru
tqdm

numpy
opencv-python
pillow

onnxruntime